addressed PR review issue: moved resize logic one logic level higher
