+++ app/src/index.css
@@ -86,3 +86,6 @@ body,
     width: 100%;
     overflow: auto;
 }
+::-webkit-scrollbar {
+    display:none;
+}
