+++ app/src/routes/project/LayersPanel/index.tsx
@@ -3,53 +3,20 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { EditorMode } from '@/lib/models';
 import clsx from 'clsx';
 import { observer } from 'mobx-react-lite';
-import { useRef, useState } from 'react';
 import { useEditorEngine } from '..';
 import LayersTab from './LayersTab';
 import { capitalizeFirstLetter } from '/common/helpers';
 import { PinLeftIcon, PinRightIcon } from '@radix-ui/react-icons';
+import { useState } from 'react';
 
 const LayersPanel = observer(() => {
     const editorEngine = useEditorEngine();
-    const panelRef = useRef<HTMLDivElement>(null);
-    const [panelWidth, setPanelWidth] = useState(240);
-    const [isOpen, setIsOpen] = useState(true);
-
-    const startResize = (e: any) => {
-        e.preventDefault();
-        e.stopPropagation();
-
-        const panel = panelRef.current;
-        if (!panel) {
-            return;
-        }
-        const startX = e.clientX;
-        const boundingRect = panel.getBoundingClientRect();
-        const startWidth = boundingRect.width;
-
-        const resize: any = (e: MouseEvent) => {
-            const currentWidth = startWidth + e.clientX - startX;
-            setPanelWidth(currentWidth);
-        };
-
-        const stopResize = (e: any) => {
-            e.preventDefault();
-            e.stopPropagation();
-
-            window.removeEventListener('mousemove', resize);
-            window.removeEventListener('mouseup', stopResize);
-        };
-
-        window.addEventListener('mousemove', resize);
-        window.addEventListener('mouseup', stopResize);
-    };
-
     enum TabValue {
         LAYERS = 'layers',
         COMPONENTS = 'components',
     }
     const selectedTab: string = TabValue.LAYERS;
-
+    const [opendrawer, setOpenDrawer] = useState(false);
     function renderTabs() {
         return (
             <Tabs defaultValue={selectedTab}>
@@ -61,7 +28,10 @@ const LayersPanel = observer(() => {
                         {capitalizeFirstLetter(TabValue.COMPONENTS)}
                     </TabsTrigger>
                     <div className="flex-grow"></div>
-                    <PinLeftIcon className="text-white cursor-pointer" onClick={() => setIsOpen(false)} />
+                    <PinLeftIcon
+                        className="text-white cursor-pointer"
+                        onClick={() => setOpenDrawer(false)}
+                    />
                 </TabsList>
                 <Separator className="mt-1" />
                 <div className="h-[calc(100vh-7.75rem)] overflow-auto mx-2">
@@ -78,25 +48,26 @@ const LayersPanel = observer(() => {
     return (
         <div
             className={clsx(
-                'fixed left-0 z-50 top-20 transition-width duration-300 opacity-100 bg-black/80',
+                'left-0 z-50 top-20 transition-width duration-300 opacity-100 bg-black/80  rounded-tr-xl',
                 editorEngine.mode === EditorMode.INTERACT ? 'hidden' : 'visible',
-                isOpen ? 'w-60 h-full rounded-tr-xl' : 'w-12 h-[5%] rounded-r-xl cursor-pointer'
+                opendrawer ? 'w-full h-full' : 'w-12 h-[5%] rounded-r-xl cursor-pointer',
             )}
-            ref={panelRef}
-            style={{ width: isOpen?`${panelWidth}px`:'' }}
         >
-            {!isOpen&&<div className='w-full h-full flex justify-center items-center cursor-pointer' onClick={()=>setIsOpen(true)}><PinRightIcon className="text-white z-51" /></div>}    
+            {!opendrawer && (
+                <div
+                    className="w-full h-full flex justify-center items-center cursor-pointer"
+                    onClick={() => setOpenDrawer(true)}
+                >
+                    <PinRightIcon className="text-white z-51" />
+                </div>
+            )}
             <div
                 className={clsx(
                     'border backdrop-blur shadow h-full relative transition-opacity duration-300 rounded-tr-xl',
-                    isOpen ? 'opacity-100 visible' : 'opacity-0 invisible'
+                    opendrawer ? 'opacity-100 visible' : 'opacity-0 invisible',
                 )}
             >
                 {renderTabs()}
-                <div
-                    className="absolute -right-1 top-0 h-full w-2 cursor-ew-resize"
-                    onMouseDown={startResize}
-                />
             </div>
         </div>
     );
